[FEAT] Fix pulseaudio plugin + add ALSA backend plugin
1. In the future commits, there will be a modular way of understanding
which sample format, sample rate, channels, etc. of the particular audio
stream coming from the server.

This should ideally be stored in a global struct of some kind and should
make it very easy for plugins to modularize their playback logic for
every variety of audio stream as supported by Wavy.

2. Also some changes to logger.hpp for better looking logs and
subsequent changes to server.

3. Added ALSA in CMakeLists (commented though)